Fix for fc1dca7 which reply_to inside Flowdock. Details follow.
Now (at least in theory) you can still respond directly inside
Flowdock to have the content reflected back in the Sprint.ly systems.
At that commit I wasn't aware than the email changes would break this
but after reading that it was a blocking point in #1, I decided to fix
it properly.